prevent git keep creation in already known foldern

+++ versioncontrol/actions/git_keep.py
@@ -1,4 +1,5 @@
 import os, platform
+from vc.apgit.repository import * 
 from git_settings import GitProjectSettings
 def on_folder_created(folder_path: str, ctx):
     # create an empty hidden .gitkeep file in the folder
@@ -8,8 +9,16 @@ def on_folder_created(folder_path: str, ctx):
     settings = GitProjectSettings(ctx)
     if not settings.gitkeep_enabled():
         return
+    
+    if ctx.project_path is None:
+        return
 
-    gitkeep_path = os.path.join(folder_path, ".gitkeep")
+    repo = GitRepository.load(ctx.project_path)
+
+    gitkeep_path = os.path.join(folder_path, ".gitkeep").replace("\\","/")
+    if repo.is_sparse_checkout_enabled() and repo.is_file_in_tree(gitkeep_path):
+         return
+
     with open(gitkeep_path, "w") as f:
         f.write("")
